Application of mopd-2.5.3-proto.patch:
* Fri Oct 26 2001 Maciej W. Rozycki <macro@ds2.pg.gda.pl>
- updated the proto, select, aout and syslog patches due to
  file location rearrangements (proto, select, aout, syslog)

* Sat Feb  5 2000 Maciej W. Rozycki <macro@ds2.pg.gda.pl>
- replaced the string patch with a more general prototype fix
  (proto)

Downloaded from:
ftp://ftp.linux-mips.org/pub/linux/mips/people/macro/mopd/mopd-2.5.3-proto.patch.gz
.patch SHA256 = 45e6a6ef14830ae32ca69b2ce5429ca86454f65e6526e86bd0db5cebf728bcd0
Patch history copied from:
ftp://ftp.linux-mips.org/pub/linux/mips/people/macro/mopd/mopd-2.5.3-15.spec

+++ common/device.c
@@ -34,6 +34,7 @@ static char rcsid[] = "$Id: device.c,v 1.21 1996/08/08 00:00:59 moj Exp $";
 #include "os.h"
 #include "common.h"
 #include "mopdef.h"
+#include "pf.h"
 
 struct	if_info *iflist;		/* Interface List		*/
 
@@ -319,7 +320,7 @@ deviceInitAll()
 	ifc.ifc_len = sizeof ibuf;
 	ifc.ifc_buf = (caddr_t)ibuf;
 	if (ioctl(fd, SIOCGIFCONF, (char *)&ifc) < 0 ||
-	    ifc.ifc_len < sizeof(struct ifreq)) {
+	    ifc.ifc_len < (int)sizeof(struct ifreq)) {
 		syslog(LOG_ERR, "deviceInitAll: old SIOCGIFCONF: %m");
 		exit(1);
 	}
